chore: move list of options to respective input component files
