+++ src/components/ElementSchemes/EventIdScheme.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
+import { optionsPropType } from '../../utils/options'
 import { SelectField } from '../'
-import { eventIdSchemeOptions } from '../../utils/options'
 
-const EventIdScheme = ({ name, label, dataTest }) => {
+const EventIdScheme = ({ name, label, eventIdSchemeOptions, dataTest }) => {
     return (
         <SelectField
             name={name}
@@ -18,6 +18,7 @@ const EventIdScheme = ({ name, label, dataTest }) => {
 
 EventIdScheme.propTypes = {
     dataTest: PropTypes.string.isRequired,
+    eventIdSchemeOptions: optionsPropType.isRequired,
     label: PropTypes.string.isRequired,
     name: PropTypes.string.isRequired,
 }
