+++ src/components/Inputs/AtomicMode.js
@@ -1,7 +1,12 @@
 import React from 'react'
 import i18n from '@dhis2/d2-i18n'
-import { atomicModeOptions } from '../../utils/options'
 import { RadioGroupField } from '../'
+
+const atomicModeOptions = [
+    { value: 'ALL', label: i18n.t('All') },
+    { value: 'NONE', label: i18n.t('None') },
+]
+const defaultAtomicModeOption = atomicModeOptions[0]
 
 const NAME = 'atomicMode'
 const DATATEST = 'input-atomic-mode'
@@ -16,4 +21,4 @@ const AtomicMode = () => (
     />
 )
 
-export { AtomicMode }
+export { AtomicMode, defaultAtomicModeOption }
