+++ src/components/Inputs/DataElementIdScheme.js
@@ -2,12 +2,24 @@ import React from 'react'
 import i18n from '@dhis2/d2-i18n'
 import { DataElementIdScheme as DataElementIdSchemeGeneric } from '../'
 
+const dataElementIdSchemeOptions = [
+    { value: 'UID', label: i18n.t('Uid') },
+    { value: 'CODE', label: i18n.t('Code') },
+    { value: 'NAME', label: i18n.t('Name') },
+]
+const defaultDataElementIdSchemeOption = dataElementIdSchemeOptions[0]
+
 const NAME = 'dataElementIdScheme'
 const DATATEST = 'input-data-element-id-scheme'
 const LABEL = i18n.t('Data element ID scheme')
 
 const DataElementIdScheme = () => (
-    <DataElementIdSchemeGeneric name={NAME} label={LABEL} dataTest={DATATEST} />
+    <DataElementIdSchemeGeneric
+        name={NAME}
+        label={LABEL}
+        dataElementIdSchemeOptions={dataElementIdSchemeOptions}
+        dataTest={DATATEST}
+    />
 )
 
-export { DataElementIdScheme }
+export { DataElementIdScheme, defaultDataElementIdSchemeOption }
