+++ src/components/Inputs/Identifier.js
@@ -1,7 +1,13 @@
 import React from 'react'
 import i18n from '@dhis2/d2-i18n'
-import { identifierOptions } from '../../utils/options'
 import { RadioGroupField } from '../'
+
+const identifierOptions = [
+    { value: 'UID', label: i18n.t('UID') },
+    { value: 'CODE', label: i18n.t('Code') },
+    { value: 'AUTO', label: i18n.t('Auto') },
+]
+const defaultIdentifierOption = identifierOptions[0]
 
 const NAME = 'identifier'
 const DATATEST = 'input-identifier'
@@ -16,4 +22,4 @@ const Identifier = () => (
     />
 )
 
-export { Identifier }
+export { Identifier, defaultIdentifierOption }
