+++ src/components/Inputs/ImportReportMode.js
@@ -1,7 +1,13 @@
 import React from 'react'
 import i18n from '@dhis2/d2-i18n'
-import { importReportModeOptions } from '../../utils/options'
 import { RadioGroupField } from '../'
+
+const importReportModeOptions = [
+    { value: 'ERRORS', label: i18n.t('Errors') },
+    { value: 'FULL', label: i18n.t('Full') },
+    { value: 'DEBUG', label: i18n.t('Debug') },
+]
+const defaultImportReportModeOption = importReportModeOptions[0]
 
 const NAME = 'importReportMode'
 const DATATEST = 'input-import-report-mode'
@@ -16,4 +22,4 @@ const ImportReportMode = () => (
     />
 )
 
-export { ImportReportMode }
+export { ImportReportMode, defaultImportReportModeOption }
