+++ src/components/Inputs/InclusionStrategy.js
@@ -1,7 +1,13 @@
 import React from 'react'
 import i18n from '@dhis2/d2-i18n'
-import { inclusionStrategyOptions } from '../../utils/options'
 import { RadioGroupField } from '../'
+
+const inclusionStrategyOptions = [
+    { value: 'NON_NULL', label: i18n.t('Non Null') },
+    { value: 'ALWAYS', label: i18n.t('Always') },
+    { value: 'NON_EMPTY', label: i18n.t('Non Empty') },
+]
+const defaultInclusionStrategyOption = inclusionStrategyOptions[0]
 
 const NAME = 'inclusionStrategy'
 const DATATEST = 'input-inclusion-strategy'
@@ -16,4 +22,4 @@ const InclusionStrategy = () => (
     />
 )
 
-export { InclusionStrategy }
+export { InclusionStrategy, defaultInclusionStrategyOption }
