+++ src/components/Inputs/LastUpdatedFilter.js
@@ -1,7 +1,13 @@
 import React from 'react'
 import i18n from '@dhis2/d2-i18n'
 import { RadioGroupField } from '../'
-import { lastUpdatedFilterOptions } from '../../utils/options'
+
+const lastUpdatedFilterOptions = [
+    { value: 'NONE', label: i18n.t('None') },
+    { value: 'DATE', label: i18n.t('Start and end date') },
+    { value: 'DURATION', label: i18n.t('Duration') },
+]
+const defaultLastUpdatedFilterOption = lastUpdatedFilterOptions[0]
 
 const NAME = 'lastUpdatedFilter'
 const DATATEST = 'input-last-updated-filter'
@@ -16,4 +22,4 @@ const LastUpdatedFilter = () => (
     />
 )
 
-export { LastUpdatedFilter }
+export { LastUpdatedFilter, defaultLastUpdatedFilterOption }
