+++ src/components/Inputs/MergeMode.js
@@ -1,7 +1,12 @@
 import React from 'react'
 import i18n from '@dhis2/d2-i18n'
-import { mergeModeOptions } from '../../utils/options'
 import { RadioGroupField } from '../'
+
+const mergeModeOptions = [
+    { value: 'MERGE', label: i18n.t('Merge') },
+    { value: 'REPLACE', label: i18n.t('Replace') },
+]
+const defaultMergeModeOption = mergeModeOptions[0]
 
 const NAME = 'mergeMode'
 const DATATEST = 'input-merge-mode'
@@ -16,4 +21,4 @@ const MergeMode = () => (
     />
 )
 
-export { MergeMode }
+export { MergeMode, defaultMergeModeOption }
