+++ src/components/Inputs/PreheatMode.js
@@ -1,7 +1,13 @@
 import React from 'react'
 import i18n from '@dhis2/d2-i18n'
-import { preheatModeOptions } from '../../utils/options'
 import { RadioGroupField } from '../'
+
+const preheatModeOptions = [
+    { value: 'REFERENCE', label: i18n.t('Reference') },
+    { value: 'ALL', label: i18n.t('All') },
+    { value: 'NONE', label: i18n.t('None') },
+]
+const defaultPreheatModeOption = preheatModeOptions[0]
 
 const NAME = 'preheatMode'
 const DATATEST = 'input-preheat-mode'
@@ -16,4 +22,4 @@ const PreheatMode = () => (
     />
 )
 
-export { PreheatMode }
+export { PreheatMode, defaultPreheatModeOption }
