+++ src/components/Inputs/Strategy.js
@@ -1,7 +1,14 @@
 import React from 'react'
 import i18n from '@dhis2/d2-i18n'
-import { strategyOptions } from '../../utils/options'
 import { RadioGroupField } from '../'
+
+const strategyOptions = [
+    { value: 'NEW_AND_UPDATES', label: i18n.t('New and updates') },
+    { value: 'NEW', label: i18n.t('New only') },
+    { value: 'UPDATES', label: i18n.t('Updates only') },
+    { value: 'DELETE', label: i18n.t('Delete') },
+]
+const defaultStrategyOption = strategyOptions[0]
 
 const NAME = 'strategy'
 const DATATEST = 'input-strategy'
@@ -16,4 +23,4 @@ const Strategy = () => (
     />
 )
 
-export { Strategy }
+export { Strategy, defaultStrategyOption }
