+++ src/components/Inputs/index.js
@@ -1,21 +1,35 @@
 export { FileUpload } from './FileUpload'
 export { ImportButtonStrip } from './ImportButtonStrip'
 export { FormAlerts } from './FormAlerts'
-export { Format } from './Format'
+export {
+    Format,
+    formatOptions,
+    formatNoCsvOptions,
+    formatAdxOptions,
+    formatAdxPdfOptions,
+    formatJsonpOptions,
+    defaultFormatOption,
+} from './Format'
 export { FirstRowIsHeader } from './FirstRowIsHeader'
 export { ClassKey } from './ClassKey'
-export { Strategy } from './Strategy'
+export { Strategy, defaultStrategyOption } from './Strategy'
 export { PreheatCache } from './PreheatCache'
 export { SkipAudit } from './SkipAudit'
 export { SkipExistingCheck } from './SkipExistingCheck'
-export { Identifier } from './Identifier'
-export { ImportReportMode } from './ImportReportMode'
-export { PreheatMode } from './PreheatMode'
-export { ImportStrategy } from './ImportStrategy'
-export { AtomicMode } from './AtomicMode'
-export { MergeMode } from './MergeMode'
-export { FlushMode } from './FlushMode'
-export { InclusionStrategy } from './InclusionStrategy'
+export { Identifier, defaultIdentifierOption } from './Identifier'
+export {
+    ImportReportMode,
+    defaultImportReportModeOption,
+} from './ImportReportMode'
+export { PreheatMode, defaultPreheatModeOption } from './PreheatMode'
+export { ImportStrategy, defaultImportStrategyOption } from './ImportStrategy'
+export { AtomicMode, defaultAtomicModeOption } from './AtomicMode'
+export { MergeMode, defaultMergeModeOption } from './MergeMode'
+export { FlushMode, defaultFlushModeOption } from './FlushMode'
+export {
+    InclusionStrategy,
+    defaultInclusionStrategyOption,
+} from './InclusionStrategy'
 export { SkipSharing } from './SkipSharing'
 export { SkipValidation } from './SkipValidation'
 export { IsAsync } from './IsAsync'
@@ -24,30 +38,42 @@ export { DataSetPicker } from './DataSetPicker'
 export { ProgramPicker } from './ProgramPicker'
 export { IncludeChildren } from './IncludeChildren'
 export { IncludeDeleted } from './IncludeDeleted'
-export { Compression } from './Compression'
-export { Inclusion } from './Inclusion'
+export { Compression, defaultCompressionOption } from './Compression'
+export { Inclusion, defaultInclusionOption } from './Inclusion'
 export { StartDate } from './StartDate'
 export { EndDate } from './EndDate'
 export { ExportButton } from './ExportButton'
 export { ProgramStages } from './ProgramStages'
 export { Schemas } from './Schemas'
-export { ObjectType } from './ObjectType'
+export { ObjectType, defaultObjectTypeOption } from './ObjectType'
 export { Objects } from './Objects'
-export { DataElementIdScheme } from './DataElementIdScheme'
-export { EventIdScheme } from './EventIdScheme'
-export { IdScheme } from './IdScheme'
-export { OrgUnitIdScheme } from './OrgUnitIdScheme'
-export { OrgUnitMode } from './OrgUnitMode'
-export { TEITypeFilter } from './TEITypeFilter'
-export { ProgramStatus } from './ProgramStatus'
-export { FollowUpStatus } from './FollowUpStatus'
+export {
+    DataElementIdScheme,
+    defaultDataElementIdSchemeOption,
+} from './DataElementIdScheme'
+export { EventIdScheme, defaultEventIdSchemeOption } from './EventIdScheme'
+export { IdScheme, defaultIdSchemeOption } from './IdScheme'
+export {
+    OrgUnitIdScheme,
+    defaultOrgUnitIdSchemeOption,
+} from './OrgUnitIdScheme'
+export { OrgUnitMode, defaultOrgUnitSelectionModeOption } from './OrgUnitMode'
+export { TEITypeFilter, defaultTEITypeFilterOption } from './TEITypeFilter'
+export { ProgramStatus, defaultProgramStatusOption } from './ProgramStatus'
+export { FollowUpStatus, defaultFollowUpStatusOption } from './FollowUpStatus'
 export { ProgramStartDate } from './ProgramStartDate'
 export { ProgramEndDate } from './ProgramEndDate'
 export { TETypePicker } from './TETypePicker'
-export { LastUpdatedFilter } from './LastUpdatedFilter'
+export {
+    LastUpdatedFilter,
+    defaultLastUpdatedFilterOption,
+} from './LastUpdatedFilter'
 export { LastUpdatedStartDate } from './LastUpdatedStartDate'
 export { LastUpdatedEndDate } from './LastUpdatedEndDate'
 export { LastUpdatedDuration } from './LastUpdatedDuration'
-export { AssignedUserMode } from './AssignedUserMode'
+export {
+    AssignedUserMode,
+    defaultAssignedUserModeOption,
+} from './AssignedUserMode'
 export { UserPicker } from './UserPicker'
 export { IncludeAllAttributes } from './IncludeAllAttributes'
