+++ src/pages/MetadataImport/MetadataImport.js
@@ -6,33 +6,31 @@ import i18n from '@dhis2/d2-i18n'
 
 import { getPrevJobDetails } from '../../utils/helper'
 import {
-    formatOptions,
-    defaultFormatOption,
-    defaultIdentifierOption,
-    defaultImportReportModeOption,
-    defaultPreheatModeOption,
-    defaultImportStrategyOption,
-    defaultAtomicModeOption,
-    defaultMergeModeOption,
-    defaultFlushModeOption,
-    defaultInclusionStrategyOption,
-} from '../../utils/options'
-import {
     FileUpload,
     Format,
+    formatOptions,
+    defaultFormatOption,
     FirstRowIsHeader,
     ClassKey,
     Identifier,
+    defaultIdentifierOption,
     ImportReportMode,
+    defaultImportReportModeOption,
     PreheatMode,
+    defaultPreheatModeOption,
     ImportStrategy,
+    defaultImportStrategyOption,
     AtomicMode,
+    defaultAtomicModeOption,
     MergeMode,
+    defaultMergeModeOption,
     FlushMode,
+    defaultFlushModeOption,
     SkipSharing,
     SkipValidation,
     IsAsync,
     InclusionStrategy,
+    defaultInclusionStrategyOption,
     ImportButtonStrip,
     FormAlerts,
 } from '../../components/Inputs'
