+++ src/pages/TEIImport/TEIImport.js
@@ -6,41 +6,39 @@ import i18n from '@dhis2/d2-i18n'
 
 import { getPrevJobDetails } from '../../utils/helper'
 import {
-    formatNoCsvOptions,
-    defaultFormatOption,
-    defaultIdentifierOption,
-    defaultImportReportModeOption,
-    defaultPreheatModeOption,
-    defaultImportStrategyOption,
-    defaultAtomicModeOption,
-    defaultMergeModeOption,
-    defaultFlushModeOption,
-    defaultInclusionStrategyOption,
-    defaultDataElementIdSchemeOption,
-    defaultOrgUnitIdSchemeOption,
-    defaultIdSchemeOption,
-    defaultEventIdSchemeOption,
-} from '../../utils/options'
-import {
     FileUpload,
     Format,
+    formatNoCsvOptions,
+    defaultFormatOption,
     Identifier,
+    defaultIdentifierOption,
     ImportReportMode,
+    defaultImportReportModeOption,
     PreheatMode,
+    defaultPreheatModeOption,
     ImportStrategy,
+    defaultImportStrategyOption,
     AtomicMode,
+    defaultAtomicModeOption,
     MergeMode,
+    defaultMergeModeOption,
     FlushMode,
+    defaultFlushModeOption,
     SkipSharing,
     SkipValidation,
     IsAsync,
     InclusionStrategy,
+    defaultInclusionStrategyOption,
     ImportButtonStrip,
     FormAlerts,
     DataElementIdScheme,
+    defaultDataElementIdSchemeOption,
     EventIdScheme,
+    defaultEventIdSchemeOption,
     IdScheme,
+    defaultIdSchemeOption,
     OrgUnitIdScheme,
+    defaultOrgUnitIdSchemeOption,
 } from '../../components/Inputs'
 import { Page, TEIIcon, MoreOptions } from '../../components/'
 import { TaskContext, getNewestTask } from '../../contexts/'
